% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{bank}
\alias{bank}
\alias{Multiserver}
\title{Bank data set}
\format{
A data frame with XXX observations on 2 variables.
\describe{
\item{arrival_time}{whats this?}
\item{service_time}{what's this?}
}
}
\usage{
data(bank)
}
\description{
Describe the dataset here in 1 sentence
The bank data frame has XXX observations on 2 variables.
}
