% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cooks_distance_scratch}
\alias{cooks_distance_scratch}
\title{Compute Cook's Distance from Scratch}
\usage{
cooks_distance_scratch(model)
}
\arguments{
\item{model}{An object of class lm.}
}
\value{
A numeric vector of Cook's distances.
}
\description{
Computes Cook's Distance for an lm model from scratch.
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
cooks_distance_scratch(model)
}
