% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dffits_scratch}
\alias{dffits_scratch}
\title{Compute DFFITS from Scratch}
\usage{
dffits_scratch(model)
}
\arguments{
\item{model}{An object of class lm.}
}
\value{
A numeric vector of DFFITS values.
}
\description{
Computes DFFITS for an lm model from scratch.
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
dffits_scratch(model)
}
