% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number indicative of the state}

\item{years}{a vector of years}
}
\value{
graphical map of the spatial distribution of fatalisties for agive state in a given year
}
\description{
This function plots the spatial distribution of fatalities corresponding to the input state number and year
}
\note{
if an invalid year or state number is defined in the input a warning message will be thrown
}
\examples{
 
fars_map_state(10,2014)

}
