% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string indicating the name of the file to read the data from}
}
\value{
this function returns a dataframe containing the data stroed in the file
}
\description{
This is a simple function to read data contained in a file
}
\note{
if the input specified file doesn't exist the function will stop and throw a file not found error
}
\examples{
  
fn<-make_filename(2013)
fars_read(fn)

}
