% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
the function returns a list of monthly data one per each year in the input vector
}
\description{
This function reads multiple data files identified by their year, one per each element of the the input vector of years
}
\note{
if an invalid year is defined in the input vector a warning message will be thrown
}
\examples{
v<-c(2014,2015) d<-fars_read_years(v)

}
