% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years}
}
\value{
a data frame with a summary per each in the input vector of years of agreggated
}
\description{
This function summarizes the data corresponding to the input vector of years with monthly aggregate
}
\note{
if an invalid year is defined in the input vector a warning message will be thrown
}
\examples{
v<-c(2014,2015) d<-fars_summarize_years(v)

}
