% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{gauss_kernel}
\alias{gauss_kernel}
\title{Gaussian kernel matrix}
\usage{
gauss_kernel(X1, X2, rho)
}
\arguments{
\item{X1}{Numeric matrix of size \eqn{n_1 \times p}.}

\item{X2}{Numeric matrix of size \eqn{n_2 \times p}.}

\item{rho}{Positive bandwidth parameter.}
}
\value{
A numeric matrix of size \eqn{n_1 \times n_2} whose (i, j)-entry is
\eqn{\exp\{-\rho \|X1_i - X2_j\|^2\}}.
}
\description{
Compute the Gaussian kernel matrix between two sets of points.
}
\examples{
X1 <- matrix(rnorm(10), ncol = 1)
X2 <- matrix(rnorm(15), ncol = 1)
K  <- gauss_kernel(X1, X2, rho = 1)
dim(K)

}
