% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hadis_influence_measure}
\alias{hadis_influence_measure}
\title{Compute Hadi's Influence Measure}
\usage{
hadis_influence_measure(model)
}
\arguments{
\item{model}{An object of class lm.}
}
\value{
A numeric vector of Hadi's influence measures.
}
\description{
Computes Hadi's Influence Measure for an lm model.
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
hadis_influence_measure(model)
}
