% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assignment2.R
\name{lat_long}
\alias{lat_long}
\title{Lat_long}
\usage{
lat_long(input, split, hemisphere)
}
\arguments{
\item{input}{Latitude or Longitude in degrees/minutes form}

\item{split}{List of delimiters between degrees, minutes, and seconds}

\item{hemisphere}{Tell lat_long if inputted values use cardinal directions to represent hemisphere}
}
\description{
This function converts Latitude and Longitude measurements in degrees/minutes/seconds form to decimal degrees form
}
