% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year:}{a string indicating the name of the file to read the data from}
}
\value{
filename corresponding to the input year
}
\description{
This is a simple function returns a filename from a year
}
\examples{
make_filename(2014)

}
