% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver for bank}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Arrivalss}

\item{ServiceTimes}{SericeTimess}

\item{NumServers}{NumServerss}
}
\value{
Bank
}
\description{
Multiserver for bank
}
\examples{
Multiserver(233,19)
}
