% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_influence}
\alias{plot_influence}
\title{Plot Influence Measures}
\usage{
plot_influence(diagnostics)
}
\arguments{
\item{diagnostics}{The result from influence_diagnostics.}
}
\value{
NULL.
}
\description{
Plots the provided influence measure results for a model.
}
\examples{
diagnostics <- influence_diagnostics(mtcars, model, "all")
plot_influence(diagnostics)
}
