% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_plot_deposit}
\alias{test_plot_deposit}
\title{Wrapper to test plot.deposit_dataset}
\usage{
test_plot_deposit(obj)
}
\arguments{
\item{obj}{Any R object}
}
\value{
A list of ggplot2 objects if successful, otherwise NULL
}
\description{
Checks that the input is of class 'deposit_dataset' before calling
the plotting method. Prints a message if the input is invalid.
}
