% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_influence}
\alias{plot_influence}
\title{Plot Influence Measures}
\usage{
plot_influence(data, model, measure = "cooks")
}
\arguments{
\item{data}{The dataset used in the model.}

\item{model}{An object of class lm.}

\item{measure}{The influence measure to plot ("cooks", "dffits", or "hadi").}
}
\description{
Plots the selected influence measure for an lm model.
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
plot_influence(mtcars, model, "cooks")
}
